Reverted straightline extraction
